✨[feat]:홈 화면 UI 구현 및 API 연동 #12
* 유저 튜표 까지 추가 구현

+++ OPeace/Projects/Core/Networking/Model/Sources/Questions/QuestionModel.swift
@@ -70,7 +70,7 @@ public struct ResultData: Codable, Equatable {
 
 // MARK: - AnswerRatio
 public struct AnswerRatio: Codable, Equatable {
-    public let a, b: Int?
+    public let a, b: Double?
 
     public enum CodingKeys: String, CodingKey {
         case a = "A"
@@ -78,8 +78,8 @@ public struct AnswerRatio: Codable, Equatable {
     }
     
     public init(
-        a: Int?,
-        b: Int?
+        a: Double?,
+        b: Double?
     ) {
         self.a = a
         self.b = b
